+++ apps/commons/src/index.tsx
@@ -1,3 +1,5 @@
+import { scan } from 'react-scan';
+
 import '@fontsource/lato/100.css';
 import '@fontsource/lato/300.css';
 import '@fontsource/lato/400.css';
@@ -6,13 +8,15 @@ import '@fontsource/lato/900.css';
 import { createOvermind } from 'overmind';
 import { Provider } from 'overmind-react';
 import { createRoot } from 'react-dom/client';
-// import { HelmetProvider } from 'react-helmet-async';
+import { HelmetProvider } from 'react-helmet-async';
 import { BrowserRouter } from 'react-router';
 import App from './App';
 import './i18n';
 import { config } from './overmind';
 import './utilities/devtoolsLog';
 import './utilities/log';
+
+scan({ enabled: true });
 
 const overmind = createOvermind(config, {
   name: 'Commons',
@@ -26,10 +30,10 @@ if (!container) throw new Error(`HTML element id 'app' not found`);
 const root = createRoot(container);
 root.render(
   <Provider value={overmind}>
-    {/* <HelmetProvider> */}
-    <BrowserRouter>
-      <App />
-    </BrowserRouter>
-    {/* </HelmetProvider> */}
+    <HelmetProvider>
+      <BrowserRouter>
+        <App />
+      </BrowserRouter>
+    </HelmetProvider>
   </Provider>,
 );