+++ packages/cwrc-leafwriter/src/dialogs/custom-authority-dialog/components/index.ts
@@ -1,0 +1,5 @@
+export * from './advanced-options';
+export * from './basic-information';
+export * from './entity-types';
+export * from './header';
+export * from './instructions';