+++ packages/cwrc-leafwriter/src/panels/markup/tree/components/TreeItem/components/Text.tsx
@@ -15,7 +15,6 @@ export const Text = forwardRef<HTMLDivElement, TextProps>(
       classnames,
       content,
       handleProps,
-      isOver,
       label,
       nodeId,
       onContextMenuOpen,